Move email modules, fix some breakage
- Moves email modules to new location

- Fixes some errors, mostly related to testing and Flask application
  context
