+++ tests/test_email.py
@@ -1,8 +1,9 @@
-from glider_dac import glider_emails, mail, app
+#from glider_dac import app
+from flask import current_app
+import glider_dac.services.emails as glider_email
 #from unittest import TestCase
 import pytest
 import smtplib
-from bson.objectid import ObjectId
 import datetime
 from unittest.mock import patch
 from types import SimpleNamespace
@@ -10,11 +11,9 @@ from types import SimpleNamespace
 @pytest.fixture
 def deployment():
     data_dict = {
-        '_id': ObjectId('000000000000000000000000'),
         'username': 'someone',
         'updated': datetime.datetime(2022, 2, 16, 21, 32, 49, 793000),
         'estimated_deploy_location': '',
-        'user_id': ObjectId('111111111111111111111111'),
         'name': 'SG610-20140715T1400',
         'archive_safe': True,
         'created': datetime.datetime(2017, 12, 4, 14, 33, 48, 513000),
@@ -42,10 +41,11 @@ def client():
    with app.test_client() as client:
        yield client
 
-@patch("glider_dac.mail")
+@patch("glider_dac.services.emails")
 def test_email_exception(client, deployment, caplog, monkeypatch):
     username = deployment.username
-    app.config["MAIL_ENABLED"] = True
+    with current_app.app_context():
+        current_app.config["MAIL_ENABLED"] = True
     def smtp_error_raiser(message):
         raise smtplib.SMTPException("Mock email failure")
     monkeypatch.setattr(mail, "send", smtp_error_raiser)
